25.09.30 hsh userController, userService 예외 작성

+++ src/main/java/com/example/demo/service/UserService.java
@@ -1,6 +1,11 @@
 package com.example.demo.service;
 
+import java.io.IOException;
+import java.util.List;
+import java.util.NoSuchElementException;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -44,7 +49,7 @@ public class UserService {
 	@Transactional
 	// 위와 같은 이유로 pet이 userId를 외래키로 가지고 있기 때문에
 	// user와 pet을 같이 받아서 처리
-	public String join(User user, Pet pet) throws Exception {
+	public String join(User user, Pet pet) throws IOException {
 
 		// Bcrypt 방식으로 암호화
 		PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
@@ -55,22 +60,25 @@ public class UserService {
 
 		// 로그인 아이디와 이메일 중복 체크
 		// 유저가 입력한 아이디가 DB에 존재하는지 확인하기 위해 select
-		// selectUserByLoginId는 DB값 있는지 검색 없으면 null 
+		// selectUserByLoginId는 DB값 있는지 검색 없으면 null
 		User dbUser = userDao.selectUserByLoginId(user.getUserLoginId()); // user.getUserLoginId()는 사용자가 입력한 아이디
 		// 이메일도 동일하게 체크
 		User dbUser2 = userDao.selectUserByEmail(user.getUserEmail());
 		// 경우에 따라서 null이라면 중복이기에 함수 종료
-		if (dbUser != null && dbUser2 != null) {
-			return "existBoth"; // 아이디 + 이메일 둘 다 중복;
-		} else if (dbUser != null) {
-			return "existID"; // 이미 존재하는 아이디
+		if (dbUser != null && dbUser2 != null) { // 아이디 + 이메일 둘 다 중복;
+			throw new DuplicateKeyException("이미 존재하는 아이디와 이메일입니다.");
+		} else if (dbUser != null) {// 이미 존재하는 아이디
+			throw new DuplicateKeyException("이미 존재하는 아이디입니다.");
 
-		} else if (dbUser2 != null) {
-			return "existEmail"; // 이미 존재하는 이메일
-
+		} else if (dbUser2 != null) { // 이미 존재하는 이메일
+			throw new DuplicateKeyException("이미 존재하는 이메일입니다.");
 		} else {
 			// 회원 정보와 반려견 정보를 DB에 저장
 			userDao.insert(user);
+			// 실패하면 spring + myBatis가 내부에서 SQL Exception 을 잡아서
+			// 적절한 런타임 에러 (ex)DataIntegrityViolationException) 을 넣어줌
+			// 그리고 GlobalExceptionHandler에서 처리함
+			// 따라서 따로 throw 안해도 됨
 
 			// Pet에 userId 세팅 후 insert
 			// Pet에 userId는 외래키이기 때문
@@ -79,6 +87,9 @@ public class UserService {
 
 			// pet 이미지 파일을 첨부 받기 위해 사용
 			MultipartFile mf = pet.getPetAttach();
+			if (mf == null) {
+				throw new IOException("파일이 제대로 첨부되지 않았습니다");
+			}
 			if (mf != null && !mf.isEmpty()) {
 				// 파일 이름 저장
 				pet.setPetAttachOname(mf.getOriginalFilename());
@@ -102,19 +113,19 @@ public class UserService {
 				dbPet.setPetAttachType(image.getPetAttachType());
 				dbPet.setPetAttachData(image.getPetAttachData());
 			}
-
 			return "success";
 		}
-
 	}
 
 	// userId로 회원 정보 조회
 	public User info(Integer userId) {
 		User user = userDao.selectByUserId(userId);
+		if (user == null)
+			throw new NoSuchElementException("해당 유저는 존재하지 않습니다");
 		return user;
 	}
 
-	// 수정 사항 중에 예외 발생 방지 위해  transactional 적용
+	// 수정 사항 중에 예외 발생 방지 위해 transactional 적용
 	@Transactional
 	public User update(User user) {
 		// db에 존재하는 회원인지 확인
@@ -122,7 +133,7 @@ public class UserService {
 
 		// 존재하지 않으면 null 반환
 		if (dbUser == null) {
-			return null;
+			throw new NoSuchElementException("해당 유저는 존재하지 않습니다");
 		}
 		// 전달된 내용이 비어있지 않으면 갱신
 		if (StringUtils.isNotBlank(user.getUserName())) {
@@ -151,14 +162,19 @@ public class UserService {
 		SUCCESS, FAIL
 	}
 
-	public RemoveResult remove(Integer userId) {
+	@Transactional
+	public User remove(Integer userId) {
+		User user = userDao.selectByUserId(userId);
+		if (user == null) {
+			throw new NoSuchElementException("해당 유저는 존재하지 않습니다");
+		}
 		// 삭제된 행 수 반환
-		int rows = userDao.delete(userId);
+		int userRows = userDao.delete(userId);
 		// 0이면 실패, 1이상이면 성공
-		if (rows == 0) {
-			return RemoveResult.FAIL;
+		if (userRows == 0) {
+			throw new NoSuchElementException("해당 유저는 존재하지 않습니다");
 		} else {
-			return RemoveResult.SUCCESS;
+			return user;
 		}
 	}
 }
